refactor: Model 패키지를 만들어 클래스 구분
