+++ src/main/java/baseball/Model/Computer.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.Model;
 
+import baseball.Constant;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
