+++ src/main/java/baseball/Model/Result.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.Model;
 
+import baseball.Constant;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
